Move test states to separate package and modules
Cloudpickle cannot work with states defined inside functions, they must be
defined in an actual module.
